fix: multisrp e2e
