+++ e2e/specs/multisrp/export-srp-from-account-actions.spec.ts
@@ -10,20 +10,12 @@ import FixtureServer from '../../fixtures/fixture-server';
 import { getFixturesServerPort } from '../../fixtures/utils';
 import { loginToApp } from '../../viewHelper';
 import TestHelpers from '../../helpers';
-import { startExportForKeyring, completeSrpQuiz } from './utils';
+import { goToAccountActions, completeSrpQuiz } from './utils';
 import { defaultOptions } from '../../seeder/anvil-manager';
-
 const fixtureServer = new FixtureServer();
 
-const SRP_1 = {
-  index: 1,
-  id: '01JN61V4CZ5WSJXSS7END4FJQ9',
-};
-
-const SRP_2 = {
-  index: 2,
-  id: '01JN61V9ACE7ZA3ZRZFPYFYCJ1',
-};
+const FIRST_DEFAULT_HD_KEYRING_ACCOUNT = 0;
+const FIRST_IMPORTED_HD_KEYRING_ACCOUNT = 2;
 
 const DEFAULT_SRP = defaultOptions.mnemonic;
 const IMPORTED_SRP =
@@ -50,12 +42,12 @@ describe(
     });
 
     it('exports the correct srp for the default hd keyring', async () => {
-      await startExportForKeyring(SRP_1.id);
+      await goToAccountActions(FIRST_DEFAULT_HD_KEYRING_ACCOUNT);
       await completeSrpQuiz(DEFAULT_SRP);
     });
 
     it('exports the correct srp for the imported hd keyring', async () => {
-      await startExportForKeyring(SRP_2.id);
+      await goToAccountActions(FIRST_IMPORTED_HD_KEYRING_ACCOUNT);
       await completeSrpQuiz(IMPORTED_SRP);
     });
   },
