+++ e2e/specs/multisrp/export-srp-from-settings.spec.ts
@@ -2,7 +2,6 @@
 import { SmokeAccounts } from '../../tags';
 import FixtureBuilder from '../../fixtures/fixture-builder';
 import {
-  defaultGanacheOptions,
   loadFixture,
   startFixtureServer,
   stopFixtureServer,
@@ -11,12 +10,20 @@ import FixtureServer from '../../fixtures/fixture-server';
 import { getFixturesServerPort } from '../../fixtures/utils';
 import { loginToApp } from '../../viewHelper';
 import TestHelpers from '../../helpers';
-import { goToAccountActions, completeSrpQuiz } from './utils';
+import { startExportForKeyring, completeSrpQuiz } from './utils';
 import { defaultOptions } from '../../seeder/anvil-manager';
+
 const fixtureServer = new FixtureServer();
 
-const FIRST_DEFAULT_HD_KEYRING_ACCOUNT = 0;
-const FIRST_IMPORTED_HD_KEYRING_ACCOUNT = 2;
+const SRP_1 = {
+  index: 1,
+  id: '01JX9NJ15HPNS6RRRYBCKDK33R',
+};
+
+const SRP_2 = {
+  index: 2,
+  id: '01JX9NZWRAVQKES02TWSN8GD91',
+};
 
 const DEFAULT_SRP = defaultOptions.mnemonic;
 const IMPORTED_SRP =
@@ -43,12 +50,12 @@ describe(
     });
 
     it('exports the correct srp for the default hd keyring', async () => {
-      await goToAccountActions(FIRST_DEFAULT_HD_KEYRING_ACCOUNT);
+      await startExportForKeyring(SRP_1.id);
       await completeSrpQuiz(DEFAULT_SRP);
     });
 
     it('exports the correct srp for the imported hd keyring', async () => {
-      await goToAccountActions(FIRST_IMPORTED_HD_KEYRING_ACCOUNT);
+      await startExportForKeyring(SRP_2.id);
       await completeSrpQuiz(IMPORTED_SRP);
     });
   },
